feat(finance): move tag-suggester from shared to finance module (PRD-212 hot-path)
PRD-212 readiness audit flagged `apps/pops-api/src/shared/tag-suggester.ts`
as a finance-owned hot path — it reads `entities` and
`transaction_tag_rules`, both finance-pillar tables. The shared/ location
was a leftover from the unified pops.db era.

- Move `shared/tag-suggester.{ts,test.ts}` to
  `modules/finance/tag-suggester/{index.ts,index.test.ts}` so the file
  sits next to its consumers (`transactions/router.ts`,
  `imports/lib/tag-management.ts`).
- Flip both finance-table reads (`transaction_tag_rules` and `entities`)
  from `getDrizzle()` to `getFinanceDrizzle()`. Corrections stay on the
  in-tree `findAllMatchingCorrections` shim, which already forwards to
  `@pops/finance-db`'s `transactionCorrectionsService`.
- Update the four call sites in pops-api (router, lib, two tests) and
  refresh two pops-finance-api docstring references to the new path.

Verified: `pnpm --filter @pops/api typecheck` clean, root
`pnpm typecheck` clean, `pnpm --filter @pops/api test src/modules/finance`
green (313 files / 4925 tests).
